[Refactor] refactoring for SVRP
